Move some funcs
